[Fix] #539 - FABController를 제거하고 컴포넌트를 TabBarController로 이동
