+++ SOPT-iOS/Projects/Features/RootFeature/Sources/ApplicationCoordinator.swift
@@ -245,13 +245,6 @@ extension ApplicationCoordinator {
             case .signIn:
                 self?.runSignInFlow(by: .rootWindow(animated: true, message: nil))
                 self?.removeDependency(coordinator)
-            case .FABMenu:
-                let vc = tabBarBuilder.makeTabBarFABMenu()
-                self?.router.present(vc, 
-                                     animated: true,
-                                     modalPresentationSytle: .overFullScreen,
-                                     modalTransitionStyle: .crossDissolve)
-                
             }
         }
