+++ SOPT-iOS/Projects/Features/TabBarFeature/Interface/Sources/TabBarBuildable.swift
@@ -11,5 +11,4 @@ import Core
 
 public protocol TabBarBuildable {
     func makeTabBar(with views: [UIViewController], userType: UserType) -> TabBarPresentable
-    func makeTabBarFABMenu() -> UIViewController
 }
