+++ SOPT-iOS/Projects/Features/TabBarFeature/Sources/Coordinator/TabBarBuilder.swift
@@ -21,9 +21,4 @@ extension TabBarBuilder: TabBarBuildable {
         let tabBarVC = TabBarController(viewModel: viewModel, tabList: views)
         return (tabBarVC, viewModel)
     }
-    
-    public func makeTabBarFABMenu() -> UIViewController {
-        let tabBarFABMenuVC = TabBarFABMenuVC()
-        return tabBarFABMenuVC
-    }
 }
