+++ SOPT-iOS/Projects/Features/TabBarFeature/Sources/ViewModel/TabBarViewModel.swift
@@ -19,7 +19,6 @@ final public class TabBarViewModel: TabBarViewModelType {
     
     public var onTabBarItemTapped: ((Int) -> Void)?
     public var showTabBarAlert: (() -> Void)?
-    public var showTabBarFABMenu: ((Bool) -> Void)?
     
     public struct Input {
         let isTabSelectedIndex: Driver<Int>
@@ -55,7 +54,6 @@ extension TabBarViewModel {
             .withUnretained(self)
             .sink { owner, _ in
                 owner.isFABTapped.toggle()
-                owner.showTabBarFABMenu?(owner.isFABTapped)
             }.store(in: cancelBag)
         
         return output
